+++ src/edu/uclouvain/core/nodus/compute/assign/workers/MSAAssignmentWorker.java
@@ -178,34 +178,40 @@ public class MSAAssignmentWorker extends AssignmentWorker {
 
           an.virtualLink.addAuxiliaryFlow(groupIndex, demand.getQuantity());
 
-          // Save this link in the detailed path table if needed
-          // if (pathWriter.isSavePaths()) {
           VirtualLink vl = an.virtualLink;
 
           switch (vl.getType()) {
             case VirtualLink.TYPE_LOAD:
-              pathCosts.ldCosts += vl.getCost(groupIndex);
+              pathCosts.ldCost += vl.getCost(groupIndex);
               pathCosts.ldDuration += vl.getDuration(groupIndex);
               loadingMode = vl.getEndVirtualNode().getMode();
               loadingMeans = vl.getEndVirtualNode().getMeans();
               break;
             case VirtualLink.TYPE_UNLOAD:
-              pathCosts.ulCosts += vl.getCost(groupIndex);
+              pathCosts.ulCost += vl.getCost(groupIndex);
               pathCosts.ulDuration += vl.getDuration(groupIndex);
               unloadingMode = vl.getBeginVirtualNode().getMode();
               unloadingMeans = vl.getBeginVirtualNode().getMeans();
               break;
             case VirtualLink.TYPE_TRANSIT:
-              pathCosts.trCosts += vl.getCost(groupIndex);
+              pathCosts.trCost += vl.getCost(groupIndex);
               pathCosts.trDuration += vl.getDuration(groupIndex);
               break;
             case VirtualLink.TYPE_TRANSHIP:
-              pathCosts.tpCosts += vl.getCost(groupIndex);
+              pathCosts.tpCost += vl.getCost(groupIndex);
               pathCosts.tpDuration += vl.getDuration(groupIndex);
               nbTranshipments++;
               break;
+            case VirtualLink.TYPE_STOP:
+              pathCosts.stpCost += vl.getCost(groupIndex);
+              pathCosts.stpDuration += vl.getDuration(groupIndex);
+              break;
+            case VirtualLink.TYPE_SWITCH:
+              pathCosts.swCost += vl.getCost(groupIndex);
+              pathCosts.swDuration += vl.getDuration(groupIndex);
+              break;
             case VirtualLink.TYPE_MOVE:
-              pathCosts.mvCosts += vl.getCost(groupIndex);
+              pathCosts.mvCost += vl.getCost(groupIndex);
               pathCosts.mvDuration += vl.getDuration(groupIndex);
               pathCosts.length += vl.getLength();
               pathWriter.savePathLink(vl);
@@ -213,7 +219,6 @@ public class MSAAssignmentWorker extends AssignmentWorker {
             default:
               break;
           }
-          // }
 
           // Go to next chunk in the path
           currentNode = predecessor;